Command line arguments should override environment variables
A previous commit added support for the PIDP_11_KNOB_ADDR and PIDP_11_KNOB_DATA environment variables. That commit caused those environment variables to override explicit -a # or -d # command line arguments, which is counter to the expected behavior.

This is a "do the right thing" commit, as the user has no control over the actual command line (it is handled in various PiDP scripts).

Thank to John Bruner for pointing this out.